==> notebooks/2.data_preprocessing.ipynb <==
Ce noteBook a pour objectif de préparer notre dataframe **best_player** avant l'entrainement de nos diffirents modèles de machine learning.

- Création de dictionnaire pour la selection des features *(columns keep & columns dummies)*
- Après plusieurs tests, nous avons décidé de ne pas standardiser les données des colonnes numériques car cela n'apporte aucune amélioration lors de l'entrainement des modèles *(cependant la possibilité de standardisation a été conservé en commentaire pour une potentielle réutilisation future)*  

import pandas as pd
import numpy as np
# from sklearn.preprocessing import StandardScaler


# columns dic
'''Columns keep
'PLAYER_HEIGHT',
'PLAYER_WEIGHT',
'PLAYER_EXP',
'GAME_PERIOD',
'GAME_PERIODE_SECOND_REMAINGING',
'SHOT_DISTANCE',
'''

columns_to_drop = [
    'PLAYER_ID',
    'PLAYER_NAME',
    'PLAYER_YEAR_START',
    'PLAYER_GAME_AGE',
    'PLAYER_BORN_YEAR',
    'GAME_ID',
    'GAME_EVENT_ID',
    'GAME_YEAR',
    'GAME_DATE',
    'GAME_SEASON_TYPE',
    'GAME_TEAM_ID',
    'GAME_TEAM_NAME',
    'GAME_HOME_TEAM',
    'GAME_AWAY_TEAM',
    'SHOT_X_LOCATION',
    'SHOT_Y_LOCATION',
    'PLAYER_BORN_YEAR'
]

# columns_to_scale = []

categorical_columns = [
    'PLAYER_POS',
    'SHOT_TYPE',
    'SHOT_ZONE_BASIC',
    'SHOT_ZONE_AREA',
    'SHOT_ZONE_RANGE'

]

def preprocess_data(data, categorical_columns, columns_to_drop, output_file):
    # Encoding categorical columns
    data_encoded = pd.get_dummies(data[categorical_columns], columns=categorical_columns)
    data_encoded.rename(columns=lambda x: x.replace(' ', '_'), inplace=True)

    # Convert boolean dummy columns to integers (0 or 1)
    boolean_columns = data_encoded.select_dtypes(include=bool).columns
    data_encoded[boolean_columns] = data_encoded[boolean_columns].astype(int)

    # Concatenate encoded data with the original DataFrame
    data = pd.concat([data, data_encoded], axis=1)

    # Drop categorical columns
    data = data.drop(categorical_columns, axis=1)

    # Drop specified columns
    data = data.drop(columns_to_drop, axis=1)
    '''
    # Scale columns
    scaler = StandardScaler()
    data[columns_to_scale] = scaler.fit_transform(data[columns_to_scale])
    '''
    # Save data
    data.to_csv(output_file, index=False)
    return data



best_player = pd.read_csv("../data/preprocessed/best_player.csv")

best_player_preprocessed = preprocess_data(
    best_player, categorical_columns, columns_to_drop, "../data/processed/best_player_preprocessed.csv"
)

display(best_player_preprocessed.head())
display(best_player_preprocessed.info())
